+++ src/cpp/session/modules/SessionCodeTools.R
@@ -1989,6 +1989,11 @@
    options <- options("repos", "download.file.method", "download.file.extra", "HTTPUserAgent")
    if (identical(options[["download.file.method"]], "curl"))
       options[["download.file.extra"]] <- .rs.downloadFileExtraWithCurlArgs()
+
+   # if there is no repo option set in the R session, read it from preferences
+   if (identical(length(options[["repos"]]), 0L)) {
+      options[["repos"]]["CRAN"] <- .rs.readUiPref("cran_mirror")$url
+   }
    
    # drop NULL entries
    options <- Filter(Negate(is.null), options)